+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Change authn-k8s to expect the client cert (passed in `X-SSL-Client-Certificate`) to be
   url-escaped.
+- Update Conjur issued certificates to use the common name derived from the authenticated
+  host, rather than use the value from the CSR.
+- Update Conjur issued certificates to include a SPIFFE SVID as a subject alternative
+  name (SAN).
 
 ## [1.2.0] - 2018-09-07
 ### Added